third_party/blink/web_tests/external/wpt/css/css-transitions/properties-value-003-expected.txt

This is a testharness.js-based test.
Found 14 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] background-image image(gradient) / values
  assert_not_equals: must not be target value after start got disallowed value "-webkit-linear-gradient(top, rgb(186, 218, 85), rgb(240, 117, 117))"
[FAIL] background-image image(gradient) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "background-image:2s" but got ""
[FAIL] background-size background-size(keyword) / values
  assert_not_equals: must not be target value after start got disallowed value "contain"
[FAIL] background-size background-size(keyword) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "background-size:2s" but got ""
[FAIL] display display(static to absolute) / values
  assert_not_equals: must not be target value after start got disallowed value "block"
[FAIL] display display(static to absolute) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "display:2s" but got ""
[FAIL] display display(block to inline-block) / values
  assert_not_equals: must not be target value after start got disallowed value "inline-block"
[FAIL] display display(block to inline-block) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "display:2s" but got ""
[FAIL] position position(static to absolute) / values
  assert_not_equals: must not be target value after start got disallowed value "absolute"
[FAIL] position position(static to absolute) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "position:2s" but got ""
[FAIL] position position(relative to absolute) / values
  assert_not_equals: must not be target value after start got disallowed value "absolute"
[FAIL] position position(relative to absolute) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "position:2s" but got ""
[FAIL] position position(absolute to fixed) / values
  assert_not_equals: must not be target value after start got disallowed value "fixed"
[FAIL] position position(absolute to fixed) / events
  assert_equals: Expected TransitionEnd events triggered on .transition expected "position:2s" but got ""
Harness: the test ran to completion.
